refactor: modularización profunda de backend, directorios y seguimiento de progreso
- **Arquitectura de Electron:** División del monolítico `main.js` extrayendo todos los eventos a la nueva carpeta `electron/ipc-handlers/`.
- **Reorganización de Servicios:** Creación de `electron/services/` para agrupar la lógica de negocio y limpieza de código muerto (`projectsDatabase.js`).
- **Scripts de Python:** Movimiento de todos los scripts (`audio_sync_analyzer.py`, `teams_converter.py`) a una nueva carpeta dedicada `/python`.
- **Corrección de Estabilidad:** Solucionado el crash (SIGABRT) en macOS durante el arranque añadiendo `await` a `checkMicrophonePermission()`.
- **Progreso de Transcripción:** Implementación de marcadores de porcentaje exactos en Python y Regex en Node.js para un seguimiento preciso (20% Preparación, 40% Micro, 40% Sistema).
- **Interfaz (UI):** Traducción al español y mapeo exacto de los pasos (1/4 a 4/4) en el componente `TranscriptionQueue.jsx`.
- **Agentic Documentation:** Descentralización de `AGENTS.md` creando documentación local en subcarpetas para optimizar el consumo de tokens de las IAs.
