+++ electron/services/audioRecorder.js
