+++ electron/services/embeddingService.js
