+++ electron/services/notificationService.js
