+++ electron/services/ragService.js
